+++ spec/truffle/options/cexts_spec.rb
@@ -1,0 +1,23 @@
+# Copyright (c) 2018 Oracle and/or its affiliates. All rights reserved. This
+# code is released under a tri EPL/GPL/LGPL license. You can use it,
+# redistribute it and/or modify it under the terms of the:
+#
+# Eclipse Public License version 1.0, or
+# GNU General Public License version 2, or
+# GNU Lesser General Public License version 2.1.
+
+require_relative '../../ruby/spec_helper'
+
+describe "The -Xcexts option" do
+  it "when disabled can run basic expressions" do
+    ruby_exe("p [1, 2, 3].map(&:succ)", options: "-Xcexts=false").should == "[2, 3, 4]\n"
+  end
+  
+  it "when disabled can use reasonable parts of the stdlib" do
+    ruby_exe("require 'uri'; p URI('http://foo.com/posts?id=30&limit=5#time=1305298413').query", options: "-Xcexts=false").should == "\"id=30&limit=5\"\n"
+  end
+  
+  it "when disabled cannot use parts of the stdlib that use C extensions" do
+    ruby_exe("require 'yaml'; p YAML.load('--- foo')", options: "-Xcexts=false", args: "2>&1").should =~ /disabled/
+  end
+end