+++ spec/truffle/options/platform_native_spec.rb
@@ -1,0 +1,25 @@
+# Copyright (c) 2018 Oracle and/or its affiliates. All rights reserved. This
+# code is released under a tri EPL/GPL/LGPL license. You can use it,
+# redistribute it and/or modify it under the terms of the:
+#
+# Eclipse Public License version 1.0, or
+# GNU General Public License version 2, or
+# GNU Lesser General Public License version 2.1.
+
+require_relative '../../ruby/spec_helper'
+
+guard -> { !Truffle::Boot.resilient_gem_home? } do
+  describe "The -Xplatform.native option" do
+    it "when disabled enables polyglot stdio" do
+      ruby_exe("p Truffle::Boot.get_option('polyglot.stdio')", options: "-Xplatform.native=false").should == "true\n"
+    end
+    
+    it "when disabled can run basic expressions" do
+      ruby_exe("p [1, 2, 3].map(&:succ)", options: "-Xplatform.native=false").should == "[2, 3, 4]\n"
+    end
+    
+    it "when disabled can use reasonable parts of the stdlib" do
+      ruby_exe("require 'uri'; p URI('http://foo.com/posts?id=30&limit=5#time=1305298413').query", options: "-Xplatform.native=false").should == "\"id=30&limit=5\"\n"
+    end
+  end
+end