+++ src/shared/java/org/truffleruby/shared/options/Options.java
@@ -47,12 +47,11 @@ public class Options {
     public final String SOURCE_ENCODING;
     public final String INTERNAL_ENCODING;
     public final String EXTERNAL_ENCODING;
-    public final boolean SINGLE_THREADED;
-    public final boolean POLYGLOT_STDIO;
-    public final boolean SYNC_STDIO;
     public final boolean NATIVE_PLATFORM;
     public final boolean NATIVE_INTERRUPT;
     public final boolean HANDLE_INTERRUPT;
+    public final boolean SINGLE_THREADED;
+    public final boolean POLYGLOT_STDIO;
     public final boolean CEXT_LOCK;
     public final boolean PREINITIALIZATION;
     public final boolean TRACE_CALLS;
@@ -133,6 +132,7 @@ public class Options {
     public final boolean SHARED_OBJECTS_DEBUG;
     public final boolean SHARED_OBJECTS_FORCE;
     public final boolean SHARED_OBJECTS_SHARE_ALL;
+    public final boolean CEXTS;
     public final boolean CEXTS_LOG_LOAD;
     public final boolean CEXTS_LOG_WARNINGS;
     public final String[] CEXTS_LIBRARY_REMAP;
@@ -175,12 +175,11 @@ public class Options {
         SOURCE_ENCODING = builder.getOrDefault(OptionsCatalog.SOURCE_ENCODING);
         INTERNAL_ENCODING = builder.getOrDefault(OptionsCatalog.INTERNAL_ENCODING);
         EXTERNAL_ENCODING = builder.getOrDefault(OptionsCatalog.EXTERNAL_ENCODING);
-        SINGLE_THREADED = builder.getOrDefault(OptionsCatalog.SINGLE_THREADED, EMBEDDED);
-        POLYGLOT_STDIO = builder.getOrDefault(OptionsCatalog.POLYGLOT_STDIO, EMBEDDED);
-        SYNC_STDIO = builder.getOrDefault(OptionsCatalog.SYNC_STDIO, EMBEDDED);
         NATIVE_PLATFORM = builder.getOrDefault(OptionsCatalog.NATIVE_PLATFORM);
         NATIVE_INTERRUPT = builder.getOrDefault(OptionsCatalog.NATIVE_INTERRUPT, NATIVE_PLATFORM);
         HANDLE_INTERRUPT = builder.getOrDefault(OptionsCatalog.HANDLE_INTERRUPT, !EMBEDDED);
+        SINGLE_THREADED = builder.getOrDefault(OptionsCatalog.SINGLE_THREADED, EMBEDDED);
+        POLYGLOT_STDIO = builder.getOrDefault(OptionsCatalog.POLYGLOT_STDIO, EMBEDDED || !NATIVE_PLATFORM);
         CEXT_LOCK = builder.getOrDefault(OptionsCatalog.CEXT_LOCK);
         PREINITIALIZATION = builder.getOrDefault(OptionsCatalog.PREINITIALIZATION);
         TRACE_CALLS = builder.getOrDefault(OptionsCatalog.TRACE_CALLS);
@@ -261,6 +260,7 @@ public class Options {
         SHARED_OBJECTS_DEBUG = builder.getOrDefault(OptionsCatalog.SHARED_OBJECTS_DEBUG);
         SHARED_OBJECTS_FORCE = builder.getOrDefault(OptionsCatalog.SHARED_OBJECTS_FORCE);
         SHARED_OBJECTS_SHARE_ALL = builder.getOrDefault(OptionsCatalog.SHARED_OBJECTS_SHARE_ALL);
+        CEXTS = builder.getOrDefault(OptionsCatalog.CEXTS);
         CEXTS_LOG_LOAD = builder.getOrDefault(OptionsCatalog.CEXTS_LOG_LOAD);
         CEXTS_LOG_WARNINGS = builder.getOrDefault(OptionsCatalog.CEXTS_LOG_WARNINGS);
         CEXTS_LIBRARY_REMAP = builder.getOrDefault(OptionsCatalog.CEXTS_LIBRARY_REMAP);
@@ -336,18 +336,16 @@ public class Options {
                 return INTERNAL_ENCODING;
             case "ruby.external_encoding":
                 return EXTERNAL_ENCODING;
-            case "ruby.single_threaded":
-                return SINGLE_THREADED;
-            case "ruby.polyglot.stdio":
-                return POLYGLOT_STDIO;
-            case "ruby.sync.stdio":
-                return SYNC_STDIO;
             case "ruby.platform.native":
                 return NATIVE_PLATFORM;
             case "ruby.platform.native_interrupt":
                 return NATIVE_INTERRUPT;
             case "ruby.platform.handle_interrupt":
                 return HANDLE_INTERRUPT;
+            case "ruby.single_threaded":
+                return SINGLE_THREADED;
+            case "ruby.polyglot.stdio":
+                return POLYGLOT_STDIO;
             case "ruby.cexts.lock":
                 return CEXT_LOCK;
             case "ruby.preinit":
@@ -508,6 +506,8 @@ public class Options {
                 return SHARED_OBJECTS_FORCE;
             case "ruby.shared.objects.share_all":
                 return SHARED_OBJECTS_SHARE_ALL;
+            case "ruby.cexts":
+                return CEXTS;
             case "ruby.cexts.log.load":
                 return CEXTS_LOG_LOAD;
             case "ruby.cexts.log.warnings":